EKIR-775 Move email helper functions to utils module
